Re structure Card components, move card styling to css file
