+++ index.html
@@ -12,7 +12,7 @@
         <p>Javascript must be enabled to use this site.</p>
     </noscript>
     <div id="app"></div>
-    <script type="text/javascript" src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDdC_USGZdqupsvV1rGpuqYi2iQQviLQwI&libraries=places"></script>
     <script type="text/javascript" src="bike-preference.1.0.0.js"></script>
+    <script type="text/javascript" src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDdC_USGZdqupsvV1rGpuqYi2iQQviLQwI&libraries=places"></script>
 </body>
 </html>
