+++ src/components/Card.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Card = props => {
+    return (
+        <div>
+            {props.children}
+        </div>
+    )
+}
+
+export default Card
