+++ src/css/index.less
@@ -1,0 +1,14 @@
+.card {
+    height: 450px;
+    width: 100%;
+    text-align: center;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.search-card {
+    .card;
+    background-image: url('http://i.imgur.com/8DfeEdo.jpg');
+    background-size: cover;
+}
